cleanup Public & Internal model
