+++ reactgrid/lib/utils/isCellSpanned.ts
@@ -1,5 +1,5 @@
-import { Cell, SpanMember } from "../types/PublicModel";
-import { SpannedCell } from "../types/InternalModel";
+import { Cell } from "../types/PublicModel";
+import { SpanMember, SpannedCell } from "../types/InternalModel";
 
 export const isCellSpanned = (cell: Cell | SpanMember): cell is SpannedCell => {
   return "rowSpan" in cell || "colSpan" in cell;
