+++ reactgrid/lib/utils/isSpanMember.ts
@@ -1,4 +1,5 @@
-import { Cell, SpanMember } from "../types/PublicModel";
+import { SpanMember } from "../types/InternalModel";
+import { Cell } from "../types/PublicModel";
 
 export const isSpanMember = (cell: Cell | SpanMember): cell is SpanMember => {
   if (!cell) return false;
