+++ reactgrid/lib/utils/scrollToElementEdge.ts
@@ -1,4 +1,4 @@
-import { Position } from "../types/PublicModel";
+import { Position } from "../types/InternalModel";
 
 export function scrollToElementEdge(relativePosition: Position, scrollableElement: HTMLElement, threshold = 0.1) {
   if (threshold <= 0) throw new Error("Scrolling threshold has to be bigger than 0.");
@@ -7,18 +7,18 @@ export function scrollToElementEdge(relativePosition: Position, scrollableElemen
 
   const touchPositionInContainer = {
     x: relativePosition.x - scrollableElementDimensions.left,
-    y: relativePosition.y - scrollableElementDimensions.top
+    y: relativePosition.y - scrollableElementDimensions.top,
   };
 
   const scrollActivationThreshold = {
     vertical: {
       min: scrollableElementDimensions.height * threshold,
-      max: scrollableElementDimensions.height - scrollableElementDimensions.height * threshold
+      max: scrollableElementDimensions.height - scrollableElementDimensions.height * threshold,
     },
     horizontal: {
       min: scrollableElementDimensions.width * threshold,
-      max: scrollableElementDimensions.width - scrollableElementDimensions.width * threshold
-    }
+      max: scrollableElementDimensions.width - scrollableElementDimensions.width * threshold,
+    },
   };
 
   const topEdgeTouched = touchPositionInContainer.y < scrollActivationThreshold.vertical.min;
