+++ reactgrid/stories/BigGrid/BigGridApi.test.tsx
@@ -8,7 +8,7 @@ import { ReactGridAPI, useReactGridAPI } from "../../lib/hooks/useReactGridAPI";
 import { TextCell } from "../../lib/cellTemplates/TextCell";
 
 describe("useReactGridAPI methods", () => {
-  let gridApi: ReactGridAPI;
+  let gridApi: ReactGridAPI | undefined;
 
   const TestComponent = () => {
     gridApi = useReactGridAPI("big-grid");
@@ -28,14 +28,14 @@ describe("useReactGridAPI methods", () => {
   });
 
   it("should set and get the selected area correctly", () => {
-    gridApi.setSelectedArea({
+    gridApi?.setSelectedArea({
       startRowIdx: 1,
       startColIdx: 3,
       endRowIdx: 1,
       endColIdx: 4,
     });
 
-    const selectedArea = gridApi.getSelectedArea();
+    const selectedArea = gridApi?.getSelectedArea();
     expect(selectedArea).toEqual({
       startRowIdx: 1,
       endRowIdx: 1,
@@ -45,9 +45,9 @@ describe("useReactGridAPI methods", () => {
   });
 
   it("should set and get the focused cell correctly", () => {
-    gridApi.setFocusedCell({ rowIndex: 5, colIndex: 8 });
+    gridApi?.setFocusedCell({ rowIndex: 5, colIndex: 8 });
 
-    const focusedCell = gridApi.getFocusedCell();
+    const focusedCell = gridApi?.getFocusedCell();
     expect(focusedCell).toEqual(
       expect.objectContaining({
         rowIndex: 5,
@@ -58,7 +58,7 @@ describe("useReactGridAPI methods", () => {
   });
 
   it("should get cell or span member by indexes correctly", () => {
-    const cell = gridApi.getCellOrSpanMemberByIndexes(1, 3);
+    const cell = gridApi?.getCellOrSpanMemberByIndexes(1, 3);
     expect(cell).toEqual(
       expect.objectContaining({
         rowIndex: 1,
@@ -69,86 +69,86 @@ describe("useReactGridAPI methods", () => {
   });
 
   it("should get pane ranges correctly", () => {
-    const paneRanges = gridApi.getPaneRanges();
+    const paneRanges = gridApi?.getPaneRanges();
     expect(paneRanges).toEqual({
       TopLeft: {
         startRowIdx: 0,
-        endRowIdx: 2,
+        endRowIdx: 1,
         startColIdx: 0,
         endColIdx: 2,
       },
       TopCenter: {
         startRowIdx: 0,
-        endRowIdx: 2,
+        endRowIdx: 1,
         startColIdx: 2,
-        endColIdx: 23,
+        endColIdx: 12,
       },
       TopRight: {
         startRowIdx: 0,
-        endRowIdx: 2,
-        startColIdx: 23,
-        endColIdx: 25,
+        endRowIdx: 1,
+        startColIdx: 12,
+        endColIdx: 14,
       },
       Left: {
-        startRowIdx: 2,
-        endRowIdx: 18,
+        startRowIdx: 1,
+        endRowIdx: 38,
         startColIdx: 0,
         endColIdx: 2,
       },
       Center: {
-        startRowIdx: 2,
-        endRowIdx: 18,
+        startRowIdx: 1,
+        endRowIdx: 38,
         startColIdx: 2,
-        endColIdx: 23,
+        endColIdx: 12,
       },
       Right: {
-        startRowIdx: 2,
-        endRowIdx: 18,
-        startColIdx: 23,
-        endColIdx: 25,
+        startRowIdx: 1,
+        endRowIdx: 38,
+        startColIdx: 12,
+        endColIdx: 14,
       },
       BottomLeft: {
-        startRowIdx: 18,
-        endRowIdx: 20,
+        startRowIdx: 38,
+        endRowIdx: 40,
         startColIdx: 0,
         endColIdx: 2,
       },
       BottomCenter: {
-        startRowIdx: 18,
-        endRowIdx: 20,
+        startRowIdx: 38,
+        endRowIdx: 40,
         startColIdx: 2,
-        endColIdx: 23,
+        endColIdx: 12,
       },
       BottomRight: {
-        startRowIdx: 18,
-        endRowIdx: 20,
-        startColIdx: 23,
-        endColIdx: 25,
+        startRowIdx: 38,
+        endRowIdx: 40,
+        startColIdx: 12,
+        endColIdx: 14,
       },
     });
   });
 
   it("should set selected columns and check selected area", () => {
-    gridApi.setSelectedColumns(2, 4);
+    gridApi?.setSelectedColumns(2, 4);
 
-    const selectedArea = gridApi.getSelectedArea();
+    const selectedArea = gridApi?.getSelectedArea();
     expect(selectedArea).toEqual({
       startRowIdx: 0,
-      endRowIdx: 20,
+      endRowIdx: 40,
       startColIdx: 2,
       endColIdx: 4,
     });
   });
 
   it("should set selected rows and check selected area", () => {
-    gridApi.setSelectedRows(1, 3);
+    gridApi?.setSelectedRows(1, 3);
 
-    const selectedArea = gridApi.getSelectedArea();
+    const selectedArea = gridApi?.getSelectedArea();
     expect(selectedArea).toEqual({
       startRowIdx: 1,
       endRowIdx: 3,
       startColIdx: 0,
-      endColIdx: 25,
+      endColIdx: 14,
     });
   });
 });
